refactor: segmenter code split to better file names
