+++ examples/segmenter/main.go
@@ -19,7 +19,6 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io"
 	"log"
 	"os"
 
@@ -81,189 +80,4 @@ func main() {
 	if err != nil {
 		log.Fatalln(err)
 	}
-}
-
-func makeSingleTrackSegments(segmenter *Segmenter, parsedMp4 *mp4.File, rs io.ReadSeeker, outFilePath string) error {
-	fileNameMap := map[string]string{"video": "_v", "audio": "_a"}
-	inits, err := segmenter.MakeInitSegments()
-	if err != nil {
-		return err
-	}
-	for _, init := range inits {
-		trackID := init.Moov.Trak.Tkhd.TrackID
-		outPath := fmt.Sprintf("%s%s%d_init.mp4", outFilePath, fileNameMap[init.GetMediaType()], trackID)
-		err = mp4.WriteToFile(init, outPath)
-		if err != nil {
-			return err
-		}
-		fmt.Printf("Generated %s\n", outPath)
-	}
-
-	segNr := 1
-	for {
-		for _, tr := range segmenter.tracks {
-			mediaType := tr.trackType
-			startSampleNr, endSampleNr := tr.segments[segNr-1].startNr, tr.segments[segNr-1].endNr
-			fmt.Printf("%s: %d-%d\n", tr.trackType, startSampleNr, endSampleNr)
-			fullSamples, err := segmenter.GetFullSamplesForInterval(parsedMp4, tr, startSampleNr, endSampleNr, rs)
-			if len(fullSamples) == 0 {
-				fmt.Printf("No more samples for %s\n", mediaType)
-				continue
-			}
-			if err != nil {
-				return err
-			}
-			seg := mp4.NewMediaSegment()
-			frag, err := mp4.CreateFragment(uint32(segNr), tr.trackID)
-			if err != nil {
-				return err
-			}
-			seg.AddFragment(frag)
-
-			for _, fullSample := range fullSamples {
-				err = frag.AddFullSampleToTrack(fullSample, tr.trackID)
-				if err != nil {
-					return err
-				}
-			}
-			outPath := fmt.Sprintf("%s%s%d_%d.m4s", outFilePath, fileNameMap[mediaType], tr.trackID, segNr)
-			err = mp4.WriteToFile(seg, outPath)
-			if err != nil {
-				return err
-			}
-			fmt.Printf("Generated %s\n", outPath)
-		}
-		segNr++
-		if segNr > segmenter.nrSegs {
-			break
-		}
-	}
-	return nil
-}
-
-func makeSingleTrackSegmentsLazyWrite(segmenter *Segmenter, parsedMp4 *mp4.File, rs io.ReadSeeker, outFilePath string) error {
-	fileNameMap := map[string]string{"video": "_v", "audio": "_a"}
-	inits, err := segmenter.MakeInitSegments()
-	if err != nil {
-		return err
-	}
-	for _, init := range inits {
-		trackID := init.Moov.Trak.Tkhd.TrackID
-		outPath := fmt.Sprintf("%s%s%d_init.mp4", outFilePath, fileNameMap[init.GetMediaType()], trackID)
-		err = mp4.WriteToFile(init, outPath)
-		if err != nil {
-			return err
-		}
-		fmt.Printf("Generated %s\n", outPath)
-	}
-
-	segNr := 1
-	for {
-		for _, tr := range segmenter.tracks {
-			mediaType := tr.trackType
-			startSampleNr, endSampleNr := tr.segments[segNr-1].startNr, tr.segments[segNr-1].endNr
-			fmt.Printf("%s: %d-%d\n", tr.trackType, startSampleNr, endSampleNr)
-			samples, err := segmenter.GetSamplesForInterval(parsedMp4, tr.inTrak, startSampleNr, endSampleNr)
-			if len(samples) == 0 {
-				fmt.Printf("No more samples for %s\n", mediaType)
-				continue
-			}
-			if err != nil {
-				return err
-			}
-			seg := mp4.NewMediaSegment()
-			frag, err := mp4.CreateFragment(uint32(segNr), tr.trackID)
-			if err != nil {
-				return err
-			}
-			seg.AddFragment(frag)
-			baseMediaDecodeTime, _ := tr.inTrak.Mdia.Minf.Stbl.Stts.GetDecodeTime(startSampleNr)
-			for _, sample := range samples {
-				err = frag.AddSampleToTrack(sample, tr.trackID, baseMediaDecodeTime)
-				if err != nil {
-					return err
-				}
-			}
-			outPath := fmt.Sprintf("%s%s%d_%d.m4s", outFilePath, fileNameMap[mediaType], tr.trackID, segNr)
-			ofh, err := os.Create(outPath)
-			if err != nil {
-				return err
-			}
-			defer ofh.Close()
-			err = seg.Encode(ofh)
-			if err != nil {
-				return err
-			}
-			// Also write media data
-			err = copyMediaData(tr.inTrak, startSampleNr, endSampleNr, rs, ofh)
-			if err != nil {
-				return err
-			}
-			fmt.Printf("Generated %s\n", outPath)
-		}
-		segNr++
-		if segNr > segmenter.nrSegs {
-			break
-		}
-	}
-	return nil
-}
-
-func makeMultiTrackSegments(segmenter *Segmenter, parsedMp4 *mp4.File, rs io.ReadSeeker, outFilePath string) error {
-	init, err := segmenter.MakeMuxedInitSegment()
-	if err != nil {
-		return err
-	}
-	outPath := fmt.Sprintf("%s_init.mp4", outFilePath)
-	err = mp4.WriteToFile(init, outPath)
-	if err != nil {
-		return err
-	}
-	fmt.Printf("Generated %s\n", outPath)
-	var trackIDs []uint32
-	for _, trak := range init.Moov.Traks {
-		trackIDs = append(trackIDs, trak.Tkhd.TrackID)
-	}
-
-	segNr := 1
-	for {
-		seg := mp4.NewMediaSegment()
-		frag, err := mp4.CreateMultiTrackFragment(uint32(segNr), trackIDs)
-		if err != nil {
-			return err
-		}
-		seg.AddFragment(frag)
-
-		for _, tr := range segmenter.tracks {
-			startSampleNr, endSampleNr := tr.segments[segNr-1].startNr, tr.segments[segNr-1].endNr
-			fmt.Printf("%s: %d-%d\n", tr.trackType, startSampleNr, endSampleNr)
-			fullSamples, err := segmenter.GetFullSamplesForInterval(parsedMp4, tr, startSampleNr, endSampleNr, rs)
-			if len(fullSamples) == 0 {
-				continue
-			}
-			if err != nil {
-				return err
-			}
-			for _, sample := range fullSamples {
-				err = frag.AddFullSampleToTrack(sample, tr.trackID)
-				if err != nil {
-					return err
-				}
-			}
-		}
-		outPath := fmt.Sprintf("%s_media_%d.m4s", outFilePath, segNr)
-		err = mp4.WriteToFile(seg, outPath)
-		if err != nil {
-			return err
-		}
-		if err != nil {
-			return err
-		}
-		fmt.Printf("Generated %s\n", outPath)
-		segNr++
-		if segNr > segmenter.nrSegs {
-			break
-		}
-	}
-	return nil
 }
